test: Get typescript working on gatsby node
