+++ gatsby-config.js
@@ -1,213 +1,7 @@
-module.exports = {
-  siteMetadata: {
-    title: `James Hush's Blog`,
-    author: {
-      name: `James Hush`,
-      summary: `Keynote speaker, coach & advisor`,
-    },
-    description: `Keynote speaker, coach & advisor.`,
-    siteUrl: `https://www.jameshush.com`,
-  },
-  plugins: [
-    `gatsby-plugin-postcss`,
-    `gatsby-plugin-image`,
-    {
-      resolve: `gatsby-source-filesystem`,
-      options: {
-        path: `${__dirname}/content/blog`,
-        name: `blog`,
-      },
-    },
-    {
-      resolve: `gatsby-source-filesystem`,
-      options: {
-        name: `images`,
-        path: `${__dirname}/src/images`,
-      },
-    },
-    {
-      resolve: `gatsby-transformer-remark`,
-      options: {
-        plugins: [
-          {
-            resolve: `gatsby-remark-images`,
-            options: {
-              maxWidth: 630,
-            },
-          },
-          {
-            resolve: "gatsby-remark-embed-youtube",
-            options: {
-              width: 800,
-              height: 400,
-            },
-          },
-          {
-            resolve: `gatsby-remark-responsive-iframe`,
-            options: {
-              wrapperStyle: `margin-bottom: 1.0725rem`,
-            },
-          },
-          {
-            resolve: "gatsby-remark-github",
-            options: {
-              marker: "GITHUB-EMBED",
-              insertEllipsisComments: true,
-              ellipsisPhrase: "...",
-              useCache: true,
-              cacheKey: "gatsby-remark-github-v1",
-              token: process.env.GITHUB_TOKEN,
-            },
-          },
-          `gatsby-remark-prismjs`,
-          `gatsby-remark-copy-linked-files`,
-          `gatsby-remark-smartypants`,
-        ],
-      },
-    },
-    `gatsby-transformer-sharp`,
-    `gatsby-plugin-sharp`,
-    {
-      resolve: `gatsby-plugin-google-gtag`,
-      options: {
-        // You can add multiple tracking ids and a pageview event will be fired for all of them.
-        trackingIds: [
-          `G-XNZS36KWBN`, // Google Analytics / GA
-        ],
-        // This object is used for configuration specific to this plugin
-        pluginConfig: {
-          // Puts tracking script in the head instead of the body
-          head: false,
-        },
-      },
-    },
-    {
-      resolve: `gatsby-plugin-feed`,
-      options: {
-        query: `
-          {
-            site {
-              siteMetadata {
-                title
-                description
-                siteUrl
-                site_url: siteUrl
-              }
-            }
-          }
-        `,
-        feeds: [
-          {
-            serialize: ({ query: { site, allMarkdownRemark } }) => {
-              return allMarkdownRemark.nodes.map(node => {
-                return Object.assign({}, node.frontmatter, {
-                  description: node.excerpt,
-                  date: node.frontmatter.date,
-                  url: site.siteMetadata.siteUrl + node.fields.slug,
-                  guid: site.siteMetadata.siteUrl + node.fields.slug,
-                  custom_elements: [{ "content:encoded": node.html }],
-                })
-              })
-            },
-            query: `
-              {
-                allMarkdownRemark(
-                  sort: { order: DESC, fields: [frontmatter___date] },
-                ) {
-                  nodes {
-                    excerpt
-                    html
-                    fields {
-                      slug
-                    }
-                    frontmatter {
-                      title
-                      date
-                    }
-                  }
-                }
-              }
-            `,
-            output: "/rss.xml",
-            title: "James Hush's Blog RSS Feed",
-          },
-        ],
-      },
-    },
-    {
-      resolve: "gatsby-plugin-sitemap",
-      options: {
-        query: `
-        {
-          site {
-            siteMetadata {
-              siteUrl
-            }
-          }
-          allSitePage {
-            nodes {
-              path
-            }
-          }
-          allMarkdownRemark {
-            nodes {
-              frontmatter {
-                date(formatString: "YYYY-MM-DD")
-              }
-              fields {
-                slug
-              }
-            }
-          }
-        }
-        `,
-        resolvePages: ({
-          allSitePage: { nodes: allPages },
-          allMarkdownRemark: { nodes: allPosts },
-        }) => {
-          const pages = allPosts.map(post => {
-            const path = post.fields.slug
-            const page = allPages.find(p => p.path === path)
-            if (page) {
-              return { path: page.path, date: post.frontmatter.date }
-            }
-          })
+// We register the TypeScript evaluator in gatsby-config so we don't need to do
+// it in any other .js file. It automatically reads TypeScript config from
+// tsconfig.json.
+require("ts-node").register()
 
-          return pages
-        },
-        serialize: ({ path, date }) => {
-          let entry = {
-            url: path,
-            changefreq: "daily",
-            priority: 0.5,
-          }
-
-          if (date) {
-            entry.priority = 0.7
-            entry.lastmod = date
-          }
-
-          return entry
-        },
-      },
-    },
-    {
-      resolve: `gatsby-plugin-manifest`,
-      options: {
-        name: `James Hush's Blog`,
-        short_name: `GatsbyJS`,
-        start_url: `/`,
-        background_color: `#fcf5e5`,
-        // This will impact how browsers show your PWA/website
-        // https://css-tricks.com/meta-theme-color-and-trickery/
-        theme_color: `#fcf5e6`,
-        display: `minimal-ui`,
-        icon: `src/images/icon.png`, // This path is relative to the root of the site.
-      },
-    },
-    `gatsby-plugin-react-helmet`,
-    // this (optional) plugin enables Progressive Web App + Offline functionality
-    // To learn more, visit: https://gatsby.dev/offline
-    `gatsby-plugin-offline`,
-  ],
-}
+// Use a TypeScript version of gatsby-config.js.
+module.exports = require("./gatsby-config.ts")
