+++ gatsby-node.ts
@@ -1,22 +1,30 @@
-const path = require(`path`)
-const { createFilePath } = require(`gatsby-source-filesystem`)
-const express = require(`express`)
+import { GatsbyNode } from "gatsby"
+import path from "path"
+
+// const path = require(`path`)
+import { createFilePath } from "gatsby-source-filesystem"
+
+import express from "express"
 
 // Serve files from `static` in development
 // Remove this when we migrate the old www.jameshush.com marketing
 // pages from static HTML to typescript in the static/ folder
-exports.onCreateDevServer = ({ app }) => {
+export const onCreateDevServer: GatsbyNode["onCreateDevServer"] = ({ app }) => {
   app.use(express.static("static"))
 }
 
-exports.createPages = async ({ graphql, actions, reporter }) => {
+export const createPages: GatsbyNode["createPages"] = async ({
+  graphql,
+  actions,
+  reporter,
+}) => {
   const { createPage } = actions
 
   // Define a template for blog post
   const blogPost = path.resolve(`./src/templates/blog-post.js`)
 
   // Get all markdown blog posts sorted by date
-  const result = await graphql(
+  const result = await graphql<GraphQL.Query.Posts>(
     `
       {
         allMarkdownRemark(
@@ -45,7 +53,7 @@ exports.createPages = async ({ graphql, actions, reporter }) => {
     return
   }
 
-  const posts = result.data.allMarkdownRemark.nodes
+  const posts = result.data?.allMarkdownRemark?.nodes ?? []
 
   // Create blog posts pages
   // But only if there's at least one markdown file found at "content/blog" (defined in gatsby-config.js)
@@ -76,7 +84,11 @@ exports.createPages = async ({ graphql, actions, reporter }) => {
   }
 }
 
-exports.onCreateNode = ({ node, actions, getNode }) => {
+export const onCreateNode: GatsbyNode["onCreateNode"] = ({
+  node,
+  actions,
+  getNode,
+}) => {
   const { createNodeField } = actions
 
   if (node.internal.type === `MarkdownRemark`) {
@@ -90,16 +102,17 @@ exports.onCreateNode = ({ node, actions, getNode }) => {
   }
 }
 
-exports.createSchemaCustomization = ({ actions }) => {
-  const { createTypes } = actions
+export const createSchemaCustomization: GatsbyNode["createSchemaCustomization"] =
+  ({ actions }) => {
+    const { createTypes } = actions
 
-  // Explicitly define the siteMetadata {} object
-  // This way those will always be defined even if removed from gatsby-config.js
+    // Explicitly define the siteMetadata {} object
+    // This way those will always be defined even if removed from gatsby-config.js
 
-  // Also explicitly define the Markdown frontmatter
-  // This way the "MarkdownRemark" queries will return `null` even when no
-  // blog posts are stored inside "content/blog" instead of returning an error
-  createTypes(`
+    // Also explicitly define the Markdown frontmatter
+    // This way the "MarkdownRemark" queries will return `null` even when no
+    // blog posts are stored inside "content/blog" instead of returning an error
+    createTypes(`
     type SiteSiteMetadata {
       author: Author
       siteUrl: String
@@ -126,4 +139,4 @@ exports.createSchemaCustomization = ({ actions }) => {
       slug: String
     }
   `)
-}
+  }
