+++ package.json
@@ -42,6 +42,7 @@
     "typescript": "^4.4.4"
   },
   "devDependencies": {
+    "@types/express": "^4.17.13",
     "@types/node": "^16.11.7",
     "@types/react": "^17.0.35",
     "@types/react-dom": "^17.0.11",
