--- src/global.d.ts
@@ -1,1 +1,0 @@
-declare const __PATH_PREFIX__: string
